# docs/swagger.yaml
basePath: /
definitions:
  httpapi.createSessionRequest:
    properties:
      calories_kcal:
        type: number
      distance_m:
        type: number
      duration_s:
        type: integer
      ended_at:
        type: string
      hr_avg:
        type: integer
      hr_max:
        type: integer
      intensity:
        type: number
      source:
        type: string
      speed_max_kmh:
        type: number
      sprints:
        type: integer
      started_at:
        type: string
    type: object
  httpapi.errorResponse:
    properties:
      error:
        type: string
    type: object
  httpapi.healthResponse:
    properties:
      database:
        type: string
      status:
        type: string
    type: object
  httpapi.updateProfileRequest:
    properties:
      display_name:
        type: string
      height_cm:
        type: integer
      preferred_position:
        type: string
      weight_kg:
        type: number
    type: object
  profile.Profile:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      height_cm:
        type: integer
      id:
        type: string
      preferred_position:
        type: string
      updated_at:
        type: string
      weight_kg:
        type: number
    type: object
  session.Session:
    properties:
      calories_kcal:
        type: number
      created_at:
        type: string
      distance_m:
        type: number
      duration_s:
        type: integer
      ended_at:
        type: string
      hr_avg:
        type: integer
      hr_max:
        type: integer
      id:
        type: string
      intensity:
        type: number
      source:
        type: string
      speed_max_kmh:
        type: number
      sprints:
        type: integer
      started_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for Rivalo (profiles, sessions, health).
  title: Rivalo API
  version: 0.1.0
paths:
  /health:
    get:
      description: Liveness probe. The database field is disabled, ok, or unreachable.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.healthResponse'
      summary: Service health
      tags:
      - health
  /v1/me:
    get:
      description: Returns the authenticated user's profile, creating a default one
        on first access.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.Profile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile fields.
      parameters:
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - profile
  /v1/sessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/session.Session'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: List my sessions
      tags:
      - sessions
    post:
      consumes:
      - application/json
      description: Stores a new sport session owned by the authenticated user.
      parameters:
      - description: Session payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.createSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/session.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a session
      tags:
      - sessions
  /v1/sessions/{id}:
    get:
      parameters:
      - description: Session id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session.Session'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a session
      tags:
      - sessions
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"